--- src/main/java/gamefaceproductions/gamefacewebsite/data/UserRole.java
@@ -1,5 +1,0 @@
-package gamefaceproductions.gamefacewebsite.data;
-
-public enum UserRole {
-    USER, ADMIN
-}
