+++ src/main/java/gamefaceproductions/gamefacewebsite/models/UserRole.java
@@ -1,0 +1,5 @@
+package gamefaceproductions.gamefacewebsite.models;
+
+public enum UserRole {
+    USER, ADMIN
+}
